requirements.txt
git+https://github.com/transcode-de/durga.git@0.2.0.dev2#egg=durga
pytest==2.7.3
pytest-cache==1.0
pytest-httpretty==0.2.0
pytest-mock=0.7.0
pytest-pythonpath==0.7
python-dateutil==2.4.2
requests==2.7.0
